Added support for middleware registration Added the ability to bulk register publishable providers Minor improvements
